Added native gap creation methods
